IMPALA-11957: Implement Regression functions: regr_slope(), regr_intercept() and regr_r2()
The linear regression functions fit an ordinary-least-squares regression line
to a set of number pairs which can be used both as aggregate and analytic functions.

regr_slope() takes two arguments of numeric type and returns the slope of the line.
regr_intercept() takes two arguments of numeric type and returns the y-intercept
of the regression line.
regr_r2() takes two arguments of numeric type and returns the coefficient
of determination (also called R-squared or goodness of fit) for the regression.
Change-Id: Iab6bd84ae3e0c02ec924c30183308123b951caa3